pyproject.toml
[project]
name = "led-matrix-zmq-control-api"
version = "0.0.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.5",
    "pydantic-settings>=2.6.1",
    "pyzmq>=26.2.0",
    "zeroconf>=0.136.0",
]
